Add li wrapper with disabled option
